Move `Audio` to `audio` package.
