+++ src/asset/core/index.js
@@ -1,5 +1,4 @@
 export * from './picture.js'
-export * from './sound.js'
 export * from './asset.js'
 export * from './handle.js'
 export * from './parser.js'
