+++ src/audio/assets/index.js
@@ -1,0 +1,1 @@
+export * from './audio.js'
