+++ src/audio/index.js
@@ -1,3 +1,4 @@
+export * from './assets/index.js'
 export * from './resources/index.js'
 export * from './typedef/index.js'
 export * from './plugin.js'
